Use the Facade Design Pattern to separate out the Game Engine.
