+++ Cm93.GameEngine/Basic/BasicGameEngine.cs
@@ -29,164 +29,42 @@ namespace Cm93.GameEngine.Basic
 {
 	public class BasicGameEngine : IGameEngine
 	{
-		private IDictionary<PlayerIndex, IList<Bid>> Bids { get; set; }
-		private Random Random { get; set; }
-
-		public ILookup<Team, Bid> TeamBids
-		{
-			get { return Bids.SelectMany(kvp => kvp.Value).ToLookup(b => b.PurchasingTeam); }
-		}
+		private PlayerBids PlayerBids { get; set; }
+		private MatchSimulator MatchSimulator { get; set; }
 
 		public BasicGameEngine()
 		{
-			Bids = new Dictionary<PlayerIndex, IList<Bid>>();
-			Random = new Random();
+			PlayerBids = new PlayerBids();
+			MatchSimulator = new MatchSimulator();
 		}
 
-		public void Play(IFixture fixture, IDictionary<int, Player> homeTeamFormation,
-			IDictionary<int, Player> awayTeamFormation, Action<double, double[,]> updateUi)
+		#region Player Bids
+
+		public ILookup<Team, Bid> TeamBids
 		{
-			for (var i = 0; i < 10; ++i)
-			{
-				var ballPositions = new double[Configuration.HeatMapDimensions.Item1, Configuration.HeatMapDimensions.Item2];
-				var homeTeamScore = homeTeamFormation.Values.Select(p => p.Rating * Random.NextDouble()).ToList();
-				var awayTeamScore = awayTeamFormation.Values.Select(p => p.Rating * Random.NextDouble()).ToList();
-
-				var round = homeTeamScore.Zip(awayTeamScore, (home, away) => (home * home) - (away * away)).Sum();
-
-				if (Configuration.PlayerTeamName != fixture.TeamHome.TeamName)
-					UpdateNpcTeams(homeTeamFormation);
-
-				if (Configuration.PlayerTeamName != fixture.TeamAway.TeamName)
-					UpdateNpcTeams(awayTeamFormation);
-
-				ColourPositionsAround(round > 0 ? homeTeamFormation.Values : awayTeamFormation.Values, ballPositions);
-
-				if (round > 0)
-					++fixture.ChancesHome;
-				else
-					++fixture.ChancesAway;
-
-				if (round > 3000)
-				{
-					++fixture.GoalsHome;
-					++fixture.TeamHome.Formation[homeTeamScore.
-						Select((value, index) => new { Index = index, Value = value }).
-						OrderByDescending(m => m.Value).
-						First().Index].Goals;
-				}
-				else if (round < -3200)
-				{
-					++fixture.GoalsAway;
-					++fixture.TeamAway.Formation[awayTeamScore.
-						Select((value, index) => new { Index = index, Value = value }).
-						OrderByDescending(m => m.Value).
-						First().Index].Goals;
-				}
-
-				if (updateUi == null)
-					continue;
-
-				fixture.Minutes += 9;
-
-				if (i == 9)
-					fixture.PlayingPeriod = PlayingPeriod.FullTime;
-				else if (i == 4)
-					fixture.PlayingPeriod = PlayingPeriod.HalfTime;
-				else if (i < 5)
-					fixture.PlayingPeriod = PlayingPeriod.FirstHalf;
-				else
-					fixture.PlayingPeriod = PlayingPeriod.SecondHalf;
-
-				var possession = homeTeamScore.Sum() / (homeTeamScore.Sum() + awayTeamScore.Sum());
-				updateUi(possession, ballPositions);
-
-				Thread.Sleep(1500);
-
-				if (i == 4)
-				{
-					fixture.PlayingPeriod = PlayingPeriod.SecondHalf;
-					updateUi(possession, null);
-					Thread.Sleep(3500);
-				}
-			}
+			get { return PlayerBids.TeamBids; }
 		}
 
-		private void ColourPositionsAround(IEnumerable<Player> players, double[,] ballPositions)
-		{
-			var coordinateList = new List<Tuple<int, int>>();
-
-			foreach (var location in players.Select(p => p.Location))
-			{
-				coordinateList.AddRange(Enumerable.Range(1, 100).Select(i =>
-					new Tuple<int, int>
-						(
-							(int) (location.X * Configuration.HeatMapDimensions.Item1) + Random.Next(-4, 4),
-							(int) (Configuration.HeatMapDimensions.Item2 - location.Y * Configuration.HeatMapDimensions.Item2) + Random.Next(-4, 4)
-						)
-					)
-				);
-			}
-
-			coordinateList.Where(t => t.Item1 > 0 && t.Item1 < 14 && t.Item2 > 0 && t.Item2 < 19).
-				Do(t => ballPositions[t.Item1, t.Item2] += 0.5d);
-		}
-
-		private void UpdateNpcTeams(IDictionary<int, Player> teamFormation)
-		{
-			teamFormation.Values.Do(p => p.Location.X = Random.NextDouble() * 0.84d);
-			teamFormation.Values.Do(p => p.Location.Y = Random.NextDouble() * 0.84d);
-		}
-
-		//	TODO: game simulation *and* player bidding in one class violates SOLID principles... fix it!
 		public void SubmitBid(Bid bid)
 		{
-			if (bid.BidAmount > bid.PurchasingTeam.Balance - TeamBids[bid.PurchasingTeam].Sum(b => b.BidAmount))
-				return;	//	Team can't afford it (including existing potentially successful bids)
-
-			if (bid.PurchasingTeam.Players.Any(p => p.Number == bid.PlayerNumber))
-				return;	//	That number is already taken
-
-			if (TeamBids[bid.PurchasingTeam].Any(b => b.PlayerNumber == bid.PlayerNumber))
-				return;	//	One team putting in two bids for the same team number
-
-			if (!Bids.ContainsKey(bid.Player.Index))
-				Bids[bid.Player.Index] = new List<Bid>();
-
-			if (Bids[bid.Player.Index].Any(b => b.PurchasingTeam == bid.PurchasingTeam))
-				return;	//	Already put in one bid, ignore subsequent bids for this week
-
-			Bids[bid.Player.Index].Add(bid);
+			PlayerBids.SubmitBid(bid);
 		}
 
 		public void ProcessTransfers()
 		{
-			foreach (var playerBidList in Bids.Values)
-			{
-				var highestBid = playerBidList.
-					Where(b => b.PurchasingTeam.Players.Count < Configuration.MaxSquadSize).
-					OrderByDescending(b => b.BidAmount).
-					First();
-
-				var player = highestBid.Player;
-
-				if (highestBid.BidAmount < player.ReleaseValue)
-					continue;	//	Winning bid isn't high enough
-
-				//	TODO: purchased players still existing in other team's match days.
-				highestBid.PurchasingTeam.Balance -= highestBid.BidAmount;
-				player.Team.Balance += highestBid.BidAmount;
-
-				player.Team.Players.Remove(player);
-				highestBid.PurchasingTeam.Players.Add(player);
-
-				player.Team = highestBid.PurchasingTeam;
-				player.Number = highestBid.PlayerNumber;
-				player.ResetPlayerIndex();
-			}
-
-			Bids.Clear();
+			PlayerBids.ProcessTransfers();
 		}
+
+		#endregion
+
+		#region Match Simulator
+
+		public void Play(IFixture fixture, IDictionary<int, Player> homeTeamFormation, IDictionary<int, Player> awayTeamFormation, Action<double, double[,]> updateUi)
+		{
+			MatchSimulator.Play(fixture, homeTeamFormation, awayTeamFormation, updateUi);
+		}
+
+		#endregion
 
 		//"Create multiple Helper classes (SOLID principles), this class collects the necessary calls from each and exposes them here to implement ISimulator."
 		public IList<IFixture> Fixtures
