+++ Cm93.GameEngine/Basic/MatchSimulator.cs
@@ -15,15 +15,120 @@
         You should have received a copy of the GNU General Public License
         along with Cm93. If not, see <http://www.gnu.org/licenses/>.
 */
+using Cm93.Model.Config;
+using Cm93.Model.Enumerations;
+using Cm93.Model.Helpers;
+using Cm93.Model.Interfaces;
+using Cm93.Model.Structures;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Threading;
 
 namespace Cm93.GameEngine.Basic
 {
 	public class MatchSimulator
 	{
+		private Random Random { get; set; }
+
+		public MatchSimulator()
+		{
+			Random = new Random();
+		}
+
+		public void Play(IFixture fixture, IDictionary<int, Player> homeTeamFormation,
+			IDictionary<int, Player> awayTeamFormation, Action<double, double[,]> updateUi)
+		{
+			for (var i = 0; i < 10; ++i)
+			{
+				var ballPositions = new double[Configuration.HeatMapDimensions.Item1, Configuration.HeatMapDimensions.Item2];
+				var homeTeamScore = homeTeamFormation.Values.Select(p => p.Rating * Random.NextDouble()).ToList();
+				var awayTeamScore = awayTeamFormation.Values.Select(p => p.Rating * Random.NextDouble()).ToList();
+
+				var round = homeTeamScore.Zip(awayTeamScore, (home, away) => (home * home) - (away * away)).Sum();
+
+				if (Configuration.PlayerTeamName != fixture.TeamHome.TeamName)
+					UpdateNpcTeams(homeTeamFormation);
+
+				if (Configuration.PlayerTeamName != fixture.TeamAway.TeamName)
+					UpdateNpcTeams(awayTeamFormation);
+
+				ColourPositionsAround(round > 0 ? homeTeamFormation.Values : awayTeamFormation.Values, ballPositions);
+
+				if (round > 0)
+					++fixture.ChancesHome;
+				else
+					++fixture.ChancesAway;
+
+				if (round > 3000)
+				{
+					++fixture.GoalsHome;
+					++fixture.TeamHome.Formation[homeTeamScore.
+						Select((value, index) => new { Index = index, Value = value }).
+						OrderByDescending(m => m.Value).
+						First().Index].Goals;
+				}
+				else if (round < -3200)
+				{
+					++fixture.GoalsAway;
+					++fixture.TeamAway.Formation[awayTeamScore.
+						Select((value, index) => new { Index = index, Value = value }).
+						OrderByDescending(m => m.Value).
+						First().Index].Goals;
+				}
+
+				if (updateUi == null)
+					continue;
+
+				fixture.Minutes += 9;
+
+				if (i == 9)
+					fixture.PlayingPeriod = PlayingPeriod.FullTime;
+				else if (i == 4)
+					fixture.PlayingPeriod = PlayingPeriod.HalfTime;
+				else if (i < 5)
+					fixture.PlayingPeriod = PlayingPeriod.FirstHalf;
+				else
+					fixture.PlayingPeriod = PlayingPeriod.SecondHalf;
+
+				var possession = homeTeamScore.Sum() / (homeTeamScore.Sum() + awayTeamScore.Sum());
+				updateUi(possession, ballPositions);
+
+				Thread.Sleep(1500);
+
+				if (i == 4)
+				{
+					fixture.PlayingPeriod = PlayingPeriod.SecondHalf;
+					updateUi(possession, null);
+					Thread.Sleep(3500);
+				}
+			}
+		}
+
+		private void ColourPositionsAround(IEnumerable<Player> players, double[,] ballPositions)
+		{
+			var coordinateList = new List<Tuple<int, int>>();
+
+			foreach (var location in players.Select(p => p.Location))
+			{
+				coordinateList.AddRange(Enumerable.Range(1, 100).Select(i =>
+					new Tuple<int, int>
+						(
+							(int) (location.X * Configuration.HeatMapDimensions.Item1) + Random.Next(-4, 4),
+							(int) (Configuration.HeatMapDimensions.Item2 - location.Y * Configuration.HeatMapDimensions.Item2) + Random.Next(-4, 4)
+						)
+					)
+				);
+			}
+
+			coordinateList.Where(t => t.Item1 > 0 && t.Item1 < 14 && t.Item2 > 0 && t.Item2 < 19).
+				Do(t => ballPositions[t.Item1, t.Item2] += 0.5d);
+		}
+
+		private void UpdateNpcTeams(IDictionary<int, Player> teamFormation)
+		{
+			teamFormation.Values.Do(p => p.Location.X = Random.NextDouble() * 0.84d);
+			teamFormation.Values.Do(p => p.Location.Y = Random.NextDouble() * 0.84d);
+		}
 	}
 }
